% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{data}
\name{dtraw}
\alias{dtraw}
\title{Dataset 2: a number of 9 traits measured on 35 Camelina DH lines.}
\format{
A \code{data.frame} with 35 observations of 9 variables.
\describe{
\item{\verb{DH lines}}{a character vector}
\item{\code{y}}{a numeric vector}
\item{\code{X1}}{a numeric vector}
\item{\code{X2}}{a numeric vector}
\item{\code{X3}}{a numeric vector}
\item{\code{X4}}{a numeric vector}
\item{\code{X5}}{a numeric vector}
\item{\code{X6}}{a numeric vector}
\item{\code{X7}}{a numeric vector}
\item{\code{X8}}{a numeric vector}
}
}
\usage{
data(dtraw)
}
\description{
Dataset 2: a number of 9 traits measured on 35 Camelina DH lines.
}
\examples{
\donttest{
library(Path.Analysis)
data(dtraw)
}
}
\keyword{datasets}
