% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{data}
\name{dtseq}
\alias{dtseq}
\title{Dataset 4: a dataframe consisting of 7 variables measured on 8 observations.}
\format{
A \code{data.frame} with 8 observations of 7 variables.
\describe{
\item{\code{Genotypes}}{a character vector}
\item{\code{YLD}}{a numeric vector}
\item{\code{DFT}}{a numeric vector}
\item{\code{FS}}{a numeric vector}
\item{\code{FV}}{a numeric vector}
\item{\code{FW}}{a numeric vector}
\item{\code{DFL}}{a numeric vector}
\item{\code{FLP}}{a numeric vector}
}
}
\usage{
data(dtseq)
}
\description{
Dataset 4: a dataframe consisting of 7 variables measured on 8 observations.
}
\examples{
\donttest{
library(Path.Analysis)
data(dtseq)
}
}
\keyword{datasets}
