% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{data}
\name{dtseqr}
\alias{dtseqr}
\title{Dataset5}
\format{
A \code{data.frame} with 24 observations of 7 variables.
\describe{
\item{\code{Genotypes}}{a character vector}
\item{\code{Rep}}{a numeric vector}
\item{\code{YLD}}{a numeric vector}
\item{\code{DFT}}{a numeric vector}
\item{\code{FS}}{a numeric vector}
\item{\code{FV}}{a numeric vector}
\item{\code{FW}}{a numeric vector}
\item{\code{DFL}}{a numeric vector}
\item{\code{FLP}}{a numeric vector}
}
}
\usage{
data(dtseqr)
}
\description{
Dataset5
}
\examples{
\donttest{
library(Path.Analysis)
data(dtseqr)
}
}
\keyword{datasets}
