% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{data}
\name{dtsimp}
\alias{dtsimp}
\title{Dataset 1: a dependent (y) and 3 independent(x1 to x3) variables.}
\format{
A \code{data.frame} with 105 observations of 4 variables.
\describe{
\item{\code{ y}}{a numeric vector}
\item{\code{x1}}{a numeric vector}
\item{\code{x2}}{a numeric vector}
\item{\code{x3}}{a numeric vector}
}
}
\usage{
data(dtsimp)
}
\description{
Dataset 1: a dependent (y) and 3 independent(x1 to x3) variables.
}
\examples{
\donttest{
library(Path.Analysis)
data(dtsimp)
}
}
\keyword{datasets}
