% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{data}
\name{heart}
\alias{heart}
\title{Dataset 6: Heart Disease data set}
\format{
A \code{data.frame} including 297 rows and 14 variables:
\describe{
\item{age}{Age in years (numerical).}
\item{sex}{Sex: 1 = male, 0 = female (logical).}
\item{\code{heart.disease}}{a numeric vector as dependent.}
\item{\code{biking}}{a numeric vector as the first independent.}
\item{\code{smoking}}{a numeric vector as the 2nd independent.}
}
}
\source{
The data set is belong to machine learning repository of UCI.
The original data set includes 303 patients with 6 NA's. After removing
missing values, it reduced into 297 patients.

\url{https://archive.ics.uci.edu/ml/datasets/Heart+Disease}
}
\usage{
data(heart)
}
\description{
A mixed variable dataset containing 14 variables of 297 patients for
their heart disease diagnosis.
}
\examples{
\donttest{
library(Path.Analysis)
data(heart)
}
}
\references{
Lichman, M. (2013). UCI machine learning repository.
}
\keyword{datasets}
