% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.plot.R
\name{network.plot}
\alias{network.plot}
\title{Network plot}
\usage{
network.plot(datap)
}
\arguments{
\item{datap}{The data set}
}
\value{
Returns an object of class \code{network_plot}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
\itemize{
\item \code{network.plot()} draws the network plot of path coefficients analysis
}
}
\details{
The \code{network.plot()} draws a correlogram and a heatmap
for data, if requested by user
}
\examples{
\donttest{
data(dtraw2)
network.plot(dtraw2)
}

}
\references{
Kuhn et al. 2022. corrr package. doi: 	<10.32614/CRAN.package.corrr>
https://github.com/tidymodels/corrr
}
\seealso{
\code{correlogram}, \code{diagram}, and \code{lavaan} packages
for drawing path diagrams.
}
\author{
{
Ali Arminian \href{mailto:abeyran@gmail.com}{abeyran@gmail.com}
}
}
