% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Path.Analysis.R
\docType{package}
\name{Path.Analysis}
\alias{Path.Analysis-package}
\alias{Path.Analysis}
\title{Path Coefficient Analysis}
\description{
\pkg{Path.Analysis} does descriptive statistics on dataset
and importantly graphical representation of data such as drawing
heatmaps, correlogram and path diagram.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abeyran/Path.Analysis}
  \item Report bugs at \url{https://github.com/abeyran/Path.Analysis/issues}
}

}
\author{
{
Ali Arminian \href{mailto:abeyran@gmail.com}{abeyran@gmail.com}
}
}
